--- a/src/basekernels/constant.jl
+++ b/src/basekernels/constant.jl
@@ -78,11 +78,11 @@ kappa(κ::ConstantKernel, x::Real) = only(κ.c) * one(x)
 metric(::ConstantKernel) = Delta()
 
 function kernelmatrix(k::ConstantKernel, x::AbstractVector)
-    return fill(only(k.c), length(x), length(x))
+    return Fill(only(k.c), length(x), length(x))
 end
 
 function kernelmatrix(k::ConstantKernel, x::AbstractVector, y::AbstractVector)
-    return fill(only(k.c), length(x), length(y))
+    return Fill(only(k.c), length(x), length(y))
 end