@theogf 2022-02-01
> Can we test the performance in some reliable way to avoid regressions?

You would need to add tests on the benchmark.jl in master first.