@willtebbutt 2022-02-01
> You would need to add tests on the benchmark.jl in master first.

Do I literally just need to add the `ConstantKernel` to line 19 of `benchmarks.jl`? Also, how is it that I trigger the benchmark suite?