@willtebbutt 2022-02-01
Or are the AD-related benchmarking tests not on master yet? They don't appear to use Zygote.