@theogf 2022-02-01
Docs fail because of https://github.com/JuliaGaussianProcesses/KernelFunctions.jl/blob/cbbf1d4c540909eff95170c7a591472fe5dbd2ae/examples/gaussian-process-priors/script.jl#L52

It looks one cannot add `I` to a `FillMatrix`