@devmotion 2022-02-01
Ah, that's an unfortunate upstream bug it seems (with an easy fix probably?). I guess we could just use a full matrix or `Eye` in the example.