@devmotion 2022-02-01
No strong objections, just a bit sad :cry: I'm looking forward to changing it to `Fill` (maybe open an issue?), it seems it would be a drastic improvement (2ms vs 37ns in the `foo` benchmark, and 2ms vs 5μs in Zygote.pullback) :slightly_smiling_face: